# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export interface Overview {
  /** Total spend in USD across all usage */
  totalCost: number;
  totalTokens: number;
  totalInputTokens: number;
  totalOutputTokens: number;
  agentCount: number;
  activeAgentCount: number;
  employeeCount: number;
  departmentCount: number;
  modelCount: number;
  runCount: number;
  avgCostPerAgent: number;
  /** @nullable */
  topDepartment: string | null;
  /** @nullable */
  topModel: string | null;
}

export interface TrendPoint {
  /** ISO date (YYYY-MM-DD) */
  date: string;
  cost: number;
  tokens: number;
}

export interface DepartmentSummary {
  id: string;
  name: string;
  cost: number;
  tokens: number;
  inputTokens: number;
  outputTokens: number;
  agentCount: number;
  employeeCount: number;
  runCount: number;
  /** Fraction of total org cost (0-1) */
  costShare: number;
}

export interface EmployeeSummary {
  id: string;
  name: string;
  role: string;
  departmentId: string;
  departmentName: string;
  cost: number;
  tokens: number;
  inputTokens: number;
  outputTokens: number;
  agentCount: number;
  runCount: number;
}

export interface ModelUsage {
  modelId: string;
  modelName: string;
  provider: string;
  cost: number;
  tokens: number;
}

export interface DepartmentDetail {
  id: string;
  name: string;
  cost: number;
  tokens: number;
  inputTokens: number;
  outputTokens: number;
  agentCount: number;
  employeeCount: number;
  runCount: number;
  costShare: number;
  employees: EmployeeSummary[];
  modelBreakdown: ModelUsage[];
}

export interface Agent {
  id: string;
  name: string;
  purpose: string;
  /** One of active, idle, archived */
  status: string;
  employeeId: string;
  employeeName: string;
  departmentId: string;
  departmentName: string;
  modelId: string;
  modelName: string;
  provider: string;
  cost: number;
  tokens: number;
  inputTokens: number;
  outputTokens: number;
  runCount: number;
  createdAt: string;
  lastActiveAt: string;
}

export interface EmployeeDetail {
  id: string;
  name: string;
  role: string;
  departmentId: string;
  departmentName: string;
  cost: number;
  tokens: number;
  inputTokens: number;
  outputTokens: number;
  agentCount: number;
  runCount: number;
  agents: Agent[];
  modelBreakdown: ModelUsage[];
}

export interface ModelSummary {
  id: string;
  name: string;
  provider: string;
  /** USD per 1M input tokens */
  inputPricePerMillion: number;
  /** USD per 1M output tokens */
  outputPricePerMillion: number;
  cost: number;
  tokens: number;
  inputTokens: number;
  outputTokens: number;
  agentCount: number;
  runCount: number;
  costShare: number;
}

export interface AgentRun {
  id: string;
  timestamp: string;
  inputTokens: number;
  outputTokens: number;
  tokens: number;
  cost: number;
}

export interface AgentDetail {
  id: string;
  name: string;
  purpose: string;
  status: string;
  employeeId: string;
  employeeName: string;
  departmentId: string;
  departmentName: string;
  modelId: string;
  modelName: string;
  provider: string;
  cost: number;
  tokens: number;
  inputTokens: number;
  outputTokens: number;
  runCount: number;
  createdAt: string;
  lastActiveAt: string;
  trends: TrendPoint[];
  recentRuns: AgentRun[];
}