# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /departments/{departmentId}:
    get:
      operationId: getDepartment
      tags: [departments]
      summary: Department detail
      description: A single department with its employees and usage aggregates.
      parameters:
        - name: departmentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Department detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepartmentDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /employees/{employeeId}:
    get:
      operationId: getEmployee
      tags: [employees]
      summary: Employee detail
      description: A single employee with the agents they launched and usage aggregates.
      parameters:
        - name: employeeId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Employee detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /agents/{agentId}:
    get:
      operationId: getAgent
      tags: [agents]
      summary: Agent detail
      description: A single agent with attribution, usage aggregates, and recent activity.
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Agent detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    TrendPoint:
      type: object
      required:
        - date
        - cost
        - tokens
      properties:
        date:
          type: string
          description: ISO date (YYYY-MM-DD)
        cost:
          type: number
        tokens:
          type: integer
...
    DepartmentDetail:
      type: object
      required:
        - id
        - name
        - cost
        - tokens
        - inputTokens
        - outputTokens
        - agentCount
        - employeeCount
        - runCount
        - costShare
        - employees
        - modelBreakdown
      properties:
        id:
          type: string
        name:
          type: string
        cost:
          type: number
        tokens:
          type: integer
        inputTokens:
          type: integer
        outputTokens:
          type: integer
        agentCount:
          type: integer
        employeeCount:
          type: integer
        runCount:
          type: integer
        costShare:
          type: number
        employees:
          type: array
          items:
            $ref: "#/components/schemas/EmployeeSummary"
        modelBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/ModelUsage"
    EmployeeSummary:
      type: object
      required:
        - id
        - name
        - role
        - departmentId
        - departmentName
        - cost
        - tokens
        - inputTokens
        - outputTokens
        - agentCount
        - runCount
      properties:
        id:
          type: string
        name:
          type: string
        role:
          type: string
        departmentId:
          type: string
        departmentName:
          type: string
        cost:
          type: number
        tokens:
          type: integer
        inputTokens:
          type: integer
        outputTokens:
          type: integer
        agentCount:
          type: integer
        runCount:
          type: integer
    EmployeeDetail:
      type: object
      required:
        - id
        - name
        - role
        - departmentId
        - departmentName
        - cost
        - tokens
        - inputTokens
        - outputTokens
        - agentCount
        - runCount
        - agents
        - modelBreakdown
      properties:
        id:
          type: string
        name:
          type: string
        role:
          type: string
        departmentId:
          type: string
        departmentName:
          type: string
        cost:
          type: number
        tokens:
          type: integer
        inputTokens:
          type: integer
        outputTokens:
          type: integer
        agentCount:
          type: integer
        runCount:
          type: integer
        agents:
          type: array
          items:
            $ref: "#/components/schemas/Agent"
        modelBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/ModelUsage"
...
    ModelUsage:
      type: object
      required:
        - modelId
        - modelName
        - provider
        - cost
        - tokens
      properties:
        modelId:
          type: string
        modelName:
          type: string
        provider:
          type: string
        cost:
          type: number
        tokens:
          type: integer
    Agent:
      type: object
      required:
        - id
        - name
        - purpose
        - status
        - employeeId
        - employeeName
        - departmentId
        - departmentName
        - modelId
        - modelName
        - provider
        - cost
        - tokens
        - inputTokens
        - outputTokens
        - runCount
        - createdAt
        - lastActiveAt
      properties:
        id:
          type: string
        name:
          type: string
        purpose:
          type: string
        status:
          type: string
          description: One of active, idle, archived
        employeeId:
          type: string
        employeeName:
          type: string
        departmentId:
          type: string
        departmentName:
          type: string
        modelId:
          type: string
        modelName:
          type: string
        provider:
          type: string
        cost:
          type: number
        tokens:
          type: integer
        inputTokens:
          type: integer
        outputTokens:
          type: integer
        runCount:
          type: integer
        createdAt:
          type: string
        lastActiveAt:
          type: string
    AgentDetail:
      type: object
      required:
        - id
        - name
        - purpose
        - status
        - employeeId
        - employeeName
        - departmentId
        - departmentName
        - modelId
        - modelName
        - provider
        - cost
        - tokens
        - inputTokens
        - outputTokens
        - runCount
        - createdAt
        - lastActiveAt
        - trends
        - recentRuns
      properties:
        id:
          type: string
        name:
          type: string
        purpose:
          type: string
        status:
          type: string
        employeeId:
          type: string
        employeeName:
          type: string
        departmentId:
          type: string
        departmentName:
          type: string
        modelId:
          type: string
        modelName:
          type: string
        provider:
          type: string
        cost:
          type: number
        tokens:
          type: integer
        inputTokens:
          type: integer
        outputTokens:
          type: integer
        runCount:
          type: integer
        createdAt:
          type: string
        lastActiveAt:
          type: string
        trends:
          type: array
          items:
            $ref: "#/components/schemas/TrendPoint"
        recentRuns:
          type: array
          items:
            $ref: "#/components/schemas/AgentRun"
    AgentRun:
      type: object
      required:
        - id
        - timestamp
        - inputTokens
        - outputTokens
        - tokens
        - cost
      properties:
        id:
          type: string
        timestamp:
          type: string
        inputTokens:
          type: integer
        outputTokens:
          type: integer
        tokens:
          type: integer
        cost:
          type: number