# lib/api-zod/src/generated/types/agent.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Agent {
  id: string;
  name: string;
  purpose: string;
  /** One of active, idle, archived */
  status: string;
  employeeId: string;
  employeeName: string;
  departmentId: string;
  departmentName: string;
  modelId: string;
  modelName: string;
  provider: string;
  cost: number;
  tokens: number;
  inputTokens: number;
  outputTokens: number;
  runCount: number;
  createdAt: string;
  lastActiveAt: string;
}