# lib/api-zod/src/generated/types/agentDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AgentRun } from './agentRun';
import type { TrendPoint } from './trendPoint';

export interface AgentDetail {
  id: string;
  name: string;
  purpose: string;
  status: string;
  employeeId: string;
  employeeName: string;
  departmentId: string;
  departmentName: string;
  modelId: string;
  modelName: string;
  provider: string;
  cost: number;
  tokens: number;
  inputTokens: number;
  outputTokens: number;
  runCount: number;
  createdAt: string;
  lastActiveAt: string;
  trends: TrendPoint[];
  recentRuns: AgentRun[];
}