# lib/api-zod/src/generated/types/agentRun.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AgentRun {
  id: string;
  timestamp: string;
  inputTokens: number;
  outputTokens: number;
  tokens: number;
  cost: number;
}