# lib/api-zod/src/generated/types/departmentDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EmployeeSummary } from './employeeSummary';
import type { ModelUsage } from './modelUsage';

export interface DepartmentDetail {
  id: string;
  name: string;
  cost: number;
  tokens: number;
  inputTokens: number;
  outputTokens: number;
  agentCount: number;
  employeeCount: number;
  runCount: number;
  costShare: number;
  employees: EmployeeSummary[];
  modelBreakdown: ModelUsage[];
}