# lib/api-zod/src/generated/types/employeeDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Agent } from './agent';
import type { ModelUsage } from './modelUsage';

export interface EmployeeDetail {
  id: string;
  name: string;
  role: string;
  departmentId: string;
  departmentName: string;
  cost: number;
  tokens: number;
  inputTokens: number;
  outputTokens: number;
  agentCount: number;
  runCount: number;
  agents: Agent[];
  modelBreakdown: ModelUsage[];
}