# lib/api-zod/src/generated/types/trendPoint.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TrendPoint {
  /** ISO date (YYYY-MM-DD) */
  date: string;
  cost: number;
  tokens: number;
}